Keep updating the UI as the program runs.
Shovelling data in from other goroutines just works and scrolling is butter smooth and this is taking neligible fractions of my CPU and RAM.  This is really impressive.

(Yes, there's a data race on my text list components.  I'm suprised it's not more troublesome; meanwhile, I'm just happy that `DataChanged` is a thing I can call, and it DTRT instead of forcing me to jump through hoops with scheduling of callbacks, like other frameworks I have known.)

Not entirely sure what the deal is with the gxui list's requirement that objects not be "equal"; that's a little odd and definitely annoying in this case, but I'm not going to worry about it too much given how pleasant every else about this is so far.

Signed-off-by: Eric Myhre <hash@exultant.us>