Start configuring prosody for local test server
Scripts assume root because linux packaging is a continuous font of joy and I can't currently be arsed to build it from source.  So, obviously, you'll want to sandbox this.  And over the long haul we'll try to make as many of our tests as possible be units that function without this.

Signed-off-by: Eric Myhre <hash@exultant.us>