Ah.  Fully qualified names are required here.
This... was apparently being interpretted as a request to send the response to that argument... which without an '@' symbol, was seen as a domain, and caused the server to try to look up another *server*... which is just miles from what want.

I'm a little irritated that errors for that aren't sent back on the original, existing connection.

So, that makes for debug messages from my local prosody server a little less crazy, but still leaves things hanging on exchanges with crypt.mn.  Reasons unknown.

Signed-off-by: Eric Myhre <hash@exultant.us>