Break handlers out into file.
Organizationally, this will ofc still be in flux for a bit, but my rough objective is to have the main dispatch method to make it Extremely Clear what message types we support at a glance.  So, nested switches: fact of life; any further handling logic without being bundled out of sight into a method: bad.

Signed-off-by: Eric Myhre <hash@exultant.us>