Commit notes on XEPs.
We need to acknowledge some set of the hundreds of them, can safely ignore others, and probably want to slightly twist the semantics of yet more (end to end acking with unique message IDs and universal sane delivery are things I consider non-optional, and I *will* break the spec to do it... just... hopefully as subtly as possible).

XMPP is a fractal surface area and I kind of hate it and want to just ignore the whole thing and start over with a sane protocol.

Signed-off-by: Eric Myhre <hash@exultant.us>