Hideously silly, abortive attempts to make tables.
Tables, it turns out, are hard, and mean different things to different people.  This was an investagatory attempt... I kinda now know how to make new components with custom paint, etc; it's *not* a fully worked do-what-i-mean replacement for anything like html tables.  If one wants that... there's a lot of questions that need to be answered up front about springiness vs absolute relations, etc etc.

Signed-off-by: Eric Myhre <hash@exultant.us>